io: forward poll_write_vectored on io::WriteHalf
Call through to the stream object to make use of the
vectored write implementation there.
Also implemented is_write_vectored by making an
is_write_vectored call to the underlying stream during split
and storing the result in the WriteHalf structure.
This assumes that the underlying stream cannot opt in and out
of efficient vectored output during its lifetime.